# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phlag API",
        "description": "API for managing feature flags, environments, and evaluations within Phlag.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List projects",
                "description": "List projects with pagination support.",
                "operationId": "listProjects",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page (1-100).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of projects.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a project",
                "description": "Store a newly created project.",
                "operationId": "createProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created.",
                        "headers": {
                            "Location": {
                                "description": "URI of the created project.",
                                "schema": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request body contained malformed JSON.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "description": "Project representation returned by the API.",
                "required": [
                    "id",
                    "key",
                    "name",
                    "created_at",
                    "updated_at",
                    "environments"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "c0a8012a-7bb2-45cb-9e7a-17a4cda9e90a"
                    },
                    "key": {
                        "type": "string",
                        "example": "checkout-service"
                    },
                    "name": {
                        "type": "string",
                        "example": "Checkout Service"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "environments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Environment"
                        }
                    }
                },
                "type": "object"
            },
            "ProjectResponse": {
                "description": "Single project response envelope.",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Project"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "description": "Pagination links keyed by relation name.",
                "type": "object",
                "additionalProperties": {
                    "type": "string",
                    "nullable": true
                }
            },
            "PaginationMeta": {
                "description": "Pagination metadata returned alongside resource collections.",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "total": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "last_page": {
                        "type": "integer",
                        "minimum": 1,
                        "nullable": true
                    }
                },
                "type": "object",
                "additionalProperties": {}
            },
            "ProjectCollection": {
                "description": "Paginated project collection response.",
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
            "Environment": {
                "description": "Environment representation.",
                "required": [
                    "id",
                    "key",
                    "name",
                    "is_default",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProjectCreateRequest": {
                "description": "Payload for creating a project.",
                "required": [
                    "key",
                    "name"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "maxLength": 64,
                        "pattern": "^[a-z0-9][a-z0-9-]*$"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorViolation": {
                "description": "Details of a single validation violation.",
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Field path that failed validation.",
                        "type": "string"
                    },
                    "message": {
                        "description": "Human-readable description of the validation error.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorEnvelope": {
                "description": "Standardized error payload returned by the Phlag API.",
                "required": [
                    "code",
                    "message",
                    "status"
                ],
                "properties": {
                    "code": {
                        "description": "Machine-readable error identifier.",
                        "type": "string",
                        "enum": [
                            "bad_request",
                            "conflict",
                            "forbidden",
                            "http_error",
                            "method_not_allowed",
                            "not_implemented",
                            "rate_limited",
                            "resource_not_found",
                            "server_error",
                            "unauthorized",
                            "validation_failed"
                        ]
                    },
                    "message": {
                        "description": "Concise, human-readable explanation of the error.",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code associated with the error.",
                        "type": "integer",
                        "format": "int32"
                    },
                    "detail": {
                        "description": "Optional detail for debugging when debug mode is enabled.",
                        "type": "string",
                        "nullable": true
                    },
                    "context": {
                        "description": "Optional key-value pairs providing additional context for the error.",
                        "type": "object",
                        "nullable": true
                    },
                    "violations": {
                        "description": "List of validation violations when the error represents validation failures.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Envelope for all error responses returned by the Phlag API.",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT issued via POST /v1/auth/token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}